refactor: Extract expression tools into separate module
- Created src/tools/expression_tools.py with 3 functions:
  - get_gtex_expression() - GTEx tissue expression data (54 human tissues)
  - get_ortholog_expression() - Model organism expression patterns
  - get_pharos_targets() - Drug target and druggability information
- Updated src/tools/__init__.py to export expression_tools
- Updated server.py to import and register expression tools
- Removed expression tool implementations from server.py (~90 lines)
- server.py now ~144 lines (down from ~234 lines)

Part of PR #6 in refactoring series (see REFACTORING_PLAN.md)
