+++ src/tools/__init__.py
@@ -4,4 +4,4 @@ Tool modules for MARRVEL-MCP.
 This package contains all the MCP tool implementations organized by category.
 """
 
-__all__ = ["gene_tools", "variant_tools", "disease_tools", "ortholog_tools"]
+__all__ = ["gene_tools", "variant_tools", "disease_tools", "ortholog_tools", "expression_tools"]
